+++ BugNet/Mod.cs
@@ -112,7 +112,7 @@ namespace BugNet
                 Name = $"Critter Cage: {getLocalizedName()}",
                 Description = "It's a critter! In a cage!",
                 texture = jaTex,
-                Category = JsonAssets.Data.ObjectData.Category_.MonsterLoot,
+                Category = JsonAssets.Data.ObjectCategory.MonsterLoot,
                 CategoryTextOverride = "Critter",
                 Price = critterId.Contains("Butterfly") ? 50 : 100,
                 ContextTags = new List<string>(new[] { "critter" }),