+++ JsonAssets/Data/BigCraftableData.cs
@@ -17,51 +17,13 @@ namespace JsonAssets.Data
         public bool ReserveNextIndex { get; set; } = false; // Deprecated
         public int ReserveExtraIndexCount { get; set; } = 0;
 
-        public class Recipe_
-        {
-            public class Ingredient
-            {
-                public object Object { get; set; }
-                public int Count { get; set; }
-            }
-
-            public string SkillUnlockName { get; set; } = null;
-            public int SkillUnlockLevel { get; set; } = -1;
-
-            public int ResultCount { get; set; } = 1;
-            public IList<Ingredient> Ingredients { get; set; } = new List<Ingredient>();
-
-            public bool IsDefault { get; set; } = false;
-            public bool CanPurchase { get; set; } = false;
-            public int PurchasePrice { get; set; }
-            public string PurchaseFrom { get; set; } = "Gus";
-            public IList<string> PurchaseRequirements { get; set; } = new List<string>();
-            public IList<PurchaseData> AdditionalPurchaseData { get; set; } = new List<PurchaseData>();
-
-            internal string GetRecipeString(BigCraftableData parent)
-            {
-                string str = "";
-                foreach (var ingredient in this.Ingredients)
-                    str += Mod.instance.ResolveObjectId(ingredient.Object) + " " + ingredient.Count + " ";
-                str = str.Substring(0, str.Length - 1);
-                str += $"/what is this for?/{parent.Id} {this.ResultCount}/true/";
-                if (this.SkillUnlockName?.Length > 0 && this.SkillUnlockLevel > 0)
-                    str += this.SkillUnlockName + " " + this.SkillUnlockLevel;
-                else
-                    str += "null";
-                if (LocalizedContentManager.CurrentLanguageCode != LocalizedContentManager.LanguageCode.en)
-                    str += "/" + parent.LocalizedName();
-                return str;
-            }
-        }
-
         public string Description { get; set; }
 
         public int Price { get; set; }
 
         public bool ProvidesLight { get; set; } = false;
 
-        public Recipe_ Recipe { get; set; }
+        public BigCraftableRecipe Recipe { get; set; }
 
         public bool CanPurchase { get; set; } = false;
         public int PurchasePrice { get; set; }