+++ JsonAssets/Data/CropBonus.cs
@@ -1,0 +1,10 @@
+namespace JsonAssets.Data
+{
+    public class CropBonus
+    {
+        public int MinimumPerHarvest { get; set; }
+        public int MaximumPerHarvest { get; set; }
+        public int MaxIncreasePerFarmLevel { get; set; }
+        public double ExtraChance { get; set; }
+    }
+}