+++ JsonAssets/Data/CropData.cs
@@ -18,13 +18,7 @@ namespace JsonAssets.Data
         public string SeedName { get; set; }
         public string SeedDescription { get; set; }
 
-        public enum CropType_
-        {
-            Normal,
-            IndoorsOnly,
-            Paddy
-        }
-        public CropType_ CropType { get; set; } = CropType_.Normal;
+        public CropType CropType { get; set; } = CropType.Normal;
 
         public IList<string> Seasons { get; set; } = new List<string>();
         public IList<int> Phases { get; set; } = new List<int>();
@@ -32,14 +26,7 @@ namespace JsonAssets.Data
         public bool HarvestWithScythe { get; set; } = false;
         public bool TrellisCrop { get; set; } = false;
         public IList<Color> Colors { get; set; } = new List<Color>();
-        public class Bonus_
-        {
-            public int MinimumPerHarvest { get; set; }
-            public int MaximumPerHarvest { get; set; }
-            public int MaxIncreasePerFarmLevel { get; set; }
-            public double ExtraChance { get; set; }
-        }
-        public Bonus_ Bonus { get; set; } = null;
+        public CropBonus Bonus { get; set; } = null;
 
         public IList<string> SeedPurchaseRequirements { get; set; } = new List<string>();
         public int SeedPurchasePrice { get; set; }