+++ JsonAssets/Data/CropType.cs
@@ -1,0 +1,9 @@
+namespace JsonAssets.Data
+{
+    public enum CropType
+    {
+        Normal,
+        IndoorsOnly,
+        Paddy
+    }
+}