+++ JsonAssets/Data/FenceBreakToolType.cs
@@ -1,0 +1,8 @@
+namespace JsonAssets.Data
+{
+    public enum FenceBreakToolType
+    {
+        Axe,
+        Pickaxe
+    }
+}