+++ JsonAssets/Data/ObjectFoodBuffs.cs
@@ -1,0 +1,17 @@
+namespace JsonAssets.Data
+{
+    public class ObjectFoodBuffs
+    {
+        public int Farming { get; set; } = 0;
+        public int Fishing { get; set; } = 0;
+        public int Mining { get; set; } = 0;
+        public int Luck { get; set; } = 0;
+        public int Foraging { get; set; } = 0;
+        public int MaxStamina { get; set; } = 0;
+        public int MagnetRadius { get; set; } = 0;
+        public int Speed { get; set; } = 0;
+        public int Defense { get; set; } = 0;
+        public int Attack { get; set; } = 0;
+        public int Duration { get; set; } = 0;
+    }
+}