+++ JsonAssets/Data/ObjectGiftTastes.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+
+namespace JsonAssets.Data
+{
+    public class ObjectGiftTastes
+    {
+        public IList<string> Love = new List<string>();
+        public IList<string> Like = new List<string>();
+        public IList<string> Neutral = new List<string>();
+        public IList<string> Dislike = new List<string>();
+        public IList<string> Hate = new List<string>();
+    }
+}