+++ JsonAssets/Data/ObjectIngredient.cs
@@ -1,0 +1,8 @@
+namespace JsonAssets.Data
+{
+    public class ObjectIngredient
+    {
+        public object Object { get; set; }
+        public int Count { get; set; }
+    }
+}