+++ JsonAssets/Data/WeaponType.cs
@@ -1,0 +1,14 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using StardewValley.Tools;
+
+namespace JsonAssets.Data
+{
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum WeaponType
+    {
+        Dagger = MeleeWeapon.dagger,
+        Club = MeleeWeapon.club,
+        Sword = MeleeWeapon.defenseSword
+    }
+}